# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SnackVeda e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List active products
      parameters:
        - in: query
          name: category
          required: false
          schema: { $ref: "#/components/schemas/ProductCategory" }
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
  /products/{slug}:
    get:
      operationId: getProductBySlug
      tags: [products]
      summary: Get a product by slug, with related products
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductDetail" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /orders/b2c:
    post:
      operationId: createB2cOrder
      tags: [orders]
      summary: Place a B2C order (auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/B2cOrderInput" }
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /orders/b2b:
    post:
      operationId: createB2bOrder
      tags: [orders]
      summary: Place a B2B order (auth required, must be approved)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/B2bOrderInput" }
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400":
          description: Invalid (MOQ or minimum order)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: B2B not approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /orders/{id}:
    get:
      operationId: getOrderById
      tags: [orders]
      summary: Get an order by id (auth required)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /account/orders:
    get:
      operationId: listMyOrders
      tags: [account]
      summary: List the current user's orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    ProductCategory:
      type: string
      enum: [healthy_chips, makhana, superpuffs]
    ProductStatus:
      type: string
      enum: [active, inactive, out_of_stock]
    OrderType:
      type: string
      enum: [b2c, b2b]
    OrderStatus:
      type: string
      enum: [pending, confirmed, dispatched, delivered, cancelled]
    PaymentMethod:
      type: string
      enum: [upi, bank_transfer, payment_link]
    PaymentStatus:
      type: string
      enum: [pending, received, failed, refunded]
...
    Address:
      type: object
      properties:
        id: { type: string }
        fullName: { type: string }
        phone: { type: string }
        line1: { type: string }
        line2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        isDefault: { type: boolean }
      required: [id, fullName, phone, line1, line2, city, state, pincode, isDefault]
    CreateAddressBody:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: string }
        line1: { type: string }
        line2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
      required: [fullName, phone, line1, city, state, pincode]
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        category: { $ref: "#/components/schemas/ProductCategory" }
        variant: { type: ["string", "null"] }
        b2cPrice: { type: number }
        b2bPrice: { type: number }
        moq: { type: integer }
        cartonQty: { type: integer }
        gstPercent: { type: number }
        hsnCode: { type: string }
        shelfLifeMonths: { type: integer }
        weightGrams: { type: integer }
        description: { type: ["string", "null"] }
        stockQty: { type: integer }
        status: { $ref: "#/components/schemas/ProductStatus" }
        sortOrder: { type: integer }
        imageUrl: { type: ["string", "null"] }
      required: [id, name, slug, category, variant, b2cPrice, b2bPrice, moq, cartonQty, gstPercent, hsnCode, shelfLifeMonths, weightGrams, description, stockQty, status, sortOrder, imageUrl]
    ProductDetail:
      type: object
      properties:
        product: { $ref: "#/components/schemas/Product" }
        related:
          type: array
          items: { $ref: "#/components/schemas/Product" }
      required: [product, related]
...
    CartItemInput:
      type: object
      properties:
        productId: { type: string }
        quantity: { type: integer, minimum: 1 }
      required: [productId, quantity]
...
    B2cOrderInput:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CartItemInput" }
        shippingAddress: { $ref: "#/components/schemas/CreateAddressBody" }
        paymentMethod: { $ref: "#/components/schemas/PaymentMethod" }
        paymentReference: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
      required: [items, shippingAddress, paymentMethod]
    B2bOrderInput:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CartItemInput" }
        shippingAddress: { $ref: "#/components/schemas/CreateAddressBody" }
        paymentMethod: { $ref: "#/components/schemas/PaymentMethod" }
        notes: { type: ["string", "null"] }
      required: [items, shippingAddress, paymentMethod]
    OrderItemView:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        productName: { type: string }
        slug: { type: string }
        category: { $ref: "#/components/schemas/ProductCategory" }
        quantity: { type: integer }
        unitPrice: { type: number }
        gstPercent: { type: number }
        gstAmount: { type: number }
        lineTotal: { type: number }
        hsnCode: { type: string }
      required: [id, productId, productName, slug, category, quantity, unitPrice, gstPercent, gstAmount, lineTotal, hsnCode]
    Order:
      type: object
      properties:
        id: { type: string }
        orderNumber: { type: string }
        userId: { type: string }
        customerName: { type: string }
        customerEmail: { type: string }
        orderType: { $ref: "#/components/schemas/OrderType" }
        status: { $ref: "#/components/schemas/OrderStatus" }
        subtotal: { type: number }
        discountAmount: { type: number }
        discountPercent: { type: number }
        gstAmount: { type: number }
        shippingCharge: { type: number }
        totalAmount: { type: number }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItemView" }
        shippingAddress:
          oneOf:
            - { $ref: "#/components/schemas/Address" }
            - { type: "null" }
        payment:
          oneOf:
            - { $ref: "#/components/schemas/PaymentView" }
            - { type: "null" }
        invoiceNumber: { type: ["string", "null"] }
      required: [id, orderNumber, userId, customerName, customerEmail, orderType, status, subtotal, discountAmount, discountPercent, gstAmount, shippingCharge, totalAmount, notes, createdAt, items, shippingAddress, payment, invoiceNumber]
    PaymentView:
      type: object
      properties:
        id: { type: string }
        paymentMethod: { $ref: "#/components/schemas/PaymentMethod" }
        paymentStatus: { $ref: "#/components/schemas/PaymentStatus" }
        amount: { type: number }
        referenceNumber: { type: ["string", "null"] }
        paidAt: { type: ["string", "null"], format: date-time }
      required: [id, paymentMethod, paymentStatus, amount, referenceNumber, paidAt]
...