# lib/api-zod/src/generated/types/b2cOrderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SnackVeda e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItemInput } from "./cartItemInput";
import type { CreateAddressBody } from "./createAddressBody";
import type { PaymentMethod } from "./paymentMethod";

export interface B2cOrderInput {
  items: CartItemInput[];
  shippingAddress: CreateAddressBody;
  paymentMethod: PaymentMethod;
  /** @nullable */
  paymentReference?: string | null;
  /** @nullable */
  notes?: string | null;
}